Bug38  - Prolems with mssing files when trying to SWARP background sutbracted images (#39)
This addresses issue #38

The underlying problem was that there were images that did not overlap any other images being used to create a tile image. Since these tiles did not overlap, there was no way to estimate what background should be subtracted, and so this file was not "processed" by BackStats (in the sense that there were no overlap fluxes for this file).

Now SwarpSetup checks to see that file exists in the appropriate directory before including it in the files to be swarped. One should be aware that at some level this is a band-aide, and that kred is not currently designed to handle tiles in which their are disconnected "islands" of files.

+++ py_progs/MefSum.py
@@ -149,8 +149,12 @@ def get_keyword(key,ext):
                 answer='N673'
             elif value.count('N708'):
                 answer='N708'
+            elif value.count('N501'):
+                answer='N501'
             elif value.count('r DECam'):
                 answer='r'
+            elif value.count('g DECam'):
+                answer='g'
             else:
                 print('Could not identify filter from ',answer)
                 answer='Unknown'
@@ -218,12 +222,13 @@ def get_mef_overview(field='LMC_c45'):
 
     print('XXX :',len(good),len(bad))
 
-    xgood=xtab[good]
-    xgood.write('goo_%s.txt' % field ,format='ascii.fixed_width_two_line',overwrite=True)
-    ztab=ascii.read('goo_%s.txt' %field)
-    os.remove('goo_%s.txt' %field)
-    ztab.sort(['FILTER','EXPTIME'])
-    ztab.write('Summary/%s_mef.tab' % field ,format='ascii.fixed_width_two_line',overwrite=True)
+    if len(good)>0:
+        xgood=xtab[good]
+        xgood.write('goo_%s.txt' % field ,format='ascii.fixed_width_two_line',overwrite=True)
+        ztab=ascii.read('goo_%s.txt' %field)
+        os.remove('goo_%s.txt' %field)
+        ztab.sort(['FILTER','EXPTIME'])
+        ztab.write('Summary/%s_mef.tab' % field ,format='ascii.fixed_width_two_line',overwrite=True)
 
     if len(bad)>0:
         xbad=xtab[bad]
